python_api/requirements.txt
TikTokApi
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
playwright==1.44.0
httpx==0.27.0